+++ src/bos/server/controllers/v1/sessiontemplate.py
@@ -31,12 +31,19 @@ import os
 
 from bos.server import redis_db_utils as dbutils
 from bos.server.models.v1_session_template import V1SessionTemplate as SessionTemplate  # noqa: E501
+from bos.server.models.v1_session_template_example_properties import V1SessionTemplateExampleProperties \
+                                                              as SessionTemplateExampleProperties  # noqa: E501
 from bos.server.utils import _canonize_xname
 from ..v2.sessiontemplates import get_v2_sessiontemplate, get_v2_sessiontemplates, delete_v2_sessiontemplate
 
 LOGGER = logging.getLogger('bos.server.controllers.v1.sessiontemplate')
 DB = dbutils.get_wrapper(db='session_templates')
 
+# These are examples returned by the /v1/sessiontemplatetemplate endpoint.
+# The intention is that a CLI user will save them to a file, edit them, and then
+# create a template via the CLI using the --file argument. When using the CLI,
+# the template name is not read in from that file -- it is provided using its own
+# argument. Because of this, this example template deliberately omits the "name" field.
 
 EXAMPLE_BOOT_SET = {
     "type": "your-boot-type",
@@ -55,8 +62,10 @@ EXAMPLE_SESSION_TEMPLATE = {
         "name_your_boot_set": EXAMPLE_BOOT_SET},
     "cfs": {
         "configuration": "desired-cfs-config"},
+    "description": "Describe your template",
     "enable_cfs": True}
 
+SESSION_TEMPLATE_TEMPLATE = SessionTemplateExampleProperties.from_dict(EXAMPLE_SESSION_TEMPLATE)
 
 def sanitize_xnames(st_json):
     """
@@ -193,7 +202,7 @@ def get_v1_sessiontemplatetemplate():
 
     Get the example session template
     """
-    return EXAMPLE_SESSION_TEMPLATE, 200
+    return SESSION_TEMPLATE_TEMPLATE, 200
 
 
 def delete_v1_sessiontemplate(session_template_id):